refactor(taxonomy): convert to tsx
