+++ src/taxonomy/data/constants.ts
@@ -7,12 +7,6 @@
  */
 export const MAX_TAXONOMY_ITEMS = 10000;
 
-/**
- * The type of a taxonomy.
- *
- * Note: the backend also has a "system" type, reserved for platform-defined
- * taxonomies. It is intentionally not selectable by users.
- */
 export enum TaxonomyType {
   Tags = 'tags',
   Competency = 'competency',
